+++ render_static/__init__.py
@@ -6,21 +6,29 @@ r"""
 /_/ |_|\___/_/ /_/\__,_/\___/_/      /____/\__/\__,_/\__/_/\___/
 
 """
-from datetime import datetime
-
 from .context import resolve_context
 from .resource import resource
 from .transpilers.defines_to_js import DefaultDefineTranspiler
 from .transpilers.enums_to_js import EnumClassWriter
 from .transpilers.urls_to_js import ClassURLWriter, SimpleURLWriter
 
-VERSION = (2, 1, 3)
+VERSION = (2, 2, 0)
 
 __title__ = "Django Render Static"
 __version__ = ".".join(str(i) for i in VERSION)
 __author__ = "Brian Kohan"
 __license__ = "MIT"
-__copyright__ = f"Copyright 2020-{datetime.now().year} Brian Kohan"
+__copyright__ = "Copyright 2020-2024 Brian Kohan"
+
+
+__all__ = [
+    "resource",
+    "resolve_context",
+    "DefaultDefineTranspiler",
+    "EnumClassWriter",
+    "ClassURLWriter",
+    "SimpleURLWriter",
+]
 
 
 class Jinja2DependencyNeeded:  # pylint: disable=R0903