+++ render_static/tests/js_tests.py
@@ -21,15 +21,12 @@ from django.test import override_settings
 from django.urls import reverse
 from django.urls.exceptions import NoReverseMatch
 from django.utils.module_loading import import_string
+
 from render_static import placeholders
 from render_static.tests import bad_pattern, defines
 from render_static.tests.enum_app.enums import DependentEnum, IndependentEnum
 from render_static.tests.enum_app.models import EnumTester
-from render_static.tests.tests import (
-    ENUM_STATIC_DIR,
-    GLOBAL_STATIC_DIR,
-    BaseTestCase,
-)
+from render_static.tests.tests import ENUM_STATIC_DIR, GLOBAL_STATIC_DIR, BaseTestCase
 from render_static.transpilers import CodeWriter
 from render_static.transpilers.enums_to_js import IGNORED_ENUMS
 from render_static.transpilers.urls_to_js import ClassURLWriter
@@ -2936,7 +2933,11 @@ class EnumComparator:
                         "strings": {
                             en.value.isoformat()
                             if isinstance(en.value, date)
-                            else str(en.value): str(en)
+                            else str(en.value): (
+                                str(getattr(en, to_string))
+                                if isinstance(to_string, str)
+                                else str(en)
+                            )
                             for en in cls
                         }
                     }
@@ -3143,6 +3144,7 @@ class EnumGeneratorTest(EnumComparator, BaseTestCase):
     def test_datetime_enum_to_javascript_param(self):
         call_command("renderstatic", "enum_app/test.js")
         from dateutil.parser import parse
+
         from render_static.tests.enum_app.defines import TimeEnum
 
         times = run_js_file(GLOBAL_STATIC_DIR / "enum_app/test.js").split()
@@ -3200,6 +3202,36 @@ class EnumGeneratorTest(EnumComparator, BaseTestCase):
         )
         self.assertNotIn("toString", get_content(ENUM_STATIC_DIR / "enum_app/test.js"))
         self.assertNotIn("this.str", get_content(ENUM_STATIC_DIR / "enum_app/test.js"))
+        self.assertNotIn(", str) {", get_content(ENUM_STATIC_DIR / "enum_app/test.js"))
+        self.assertNotIn(', "1");', get_content(ENUM_STATIC_DIR / "enum_app/test.js"))
+
+    @override_settings(
+        STATIC_TEMPLATES={
+            "context": {
+                "include_properties": True,
+                "class_properties": False,
+                "properties": True,
+                "symmetric_properties": False,
+                "to_string": "uri",
+            },
+            "templates": [
+                ("enum_app/test.js", {"context": {"enums": EnumTester.MapBoxStyle}}),
+            ],
+        }
+    )
+    def test_to_string_is_prop(self):
+        call_command("renderstatic", "enum_app/test.js")
+        self.enums_compare(
+            js_file=ENUM_STATIC_DIR / "enum_app/test.js",
+            enum_classes=[EnumTester.MapBoxStyle],
+            class_properties=False,
+            to_string="uri",
+        )
+        self.assertNotIn(
+            "class_name", get_content(ENUM_STATIC_DIR / "enum_app/test.js")
+        )
+        self.assertIn("toString", get_content(ENUM_STATIC_DIR / "enum_app/test.js"))
+        self.assertIn("this.uri", get_content(ENUM_STATIC_DIR / "enum_app/test.js"))
         self.assertNotIn(", str) {", get_content(ENUM_STATIC_DIR / "enum_app/test.js"))
         self.assertNotIn(', "1");', get_content(ENUM_STATIC_DIR / "enum_app/test.js"))