+++ render_static/tests/traverse_tests.py
@@ -6,6 +6,7 @@ from enum import Enum, auto
 from types import ModuleType
 
 from django.test import TestCase
+
 from render_static.transpilers import Transpiler
 from render_static.transpilers.enums_to_js import IGNORED_ENUMS
 
@@ -211,10 +212,7 @@ class TranspileTraverseTests(TestCase):
 
     def test_module_traversal(self):
         from render_static.tests.traverse import module1
-        from render_static.tests.traverse.module1 import (
-            Class1Module1,
-            Class2Module1,
-        )
+        from render_static.tests.traverse.module1 import Class1Module1, Class2Module1
 
         transpiler = TranspilerTester()
         transpiler.transpile([module1])
@@ -275,10 +273,7 @@ class TranspileTraverseTests(TestCase):
 
     def test_module_import_string_traversal(self):
         from render_static.tests.traverse import module1
-        from render_static.tests.traverse.module1 import (
-            Class1Module1,
-            Class2Module1,
-        )
+        from render_static.tests.traverse.module1 import Class1Module1, Class2Module1
 
         transpiler = TranspilerTester()
         transpiler.transpile(["render_static.tests.traverse.module1"])
@@ -339,10 +334,7 @@ class TranspileTraverseTests(TestCase):
 
     def test_deduplicate_module(self):
         from render_static.tests.traverse import module1
-        from render_static.tests.traverse.module1 import (
-            Class1Module1,
-            Class2Module1,
-        )
+        from render_static.tests.traverse.module1 import Class1Module1, Class2Module1
 
         transpiler = TranspilerTester()
         transpiler.transpile(["render_static.tests.traverse.module1", module1])
@@ -403,10 +395,7 @@ class TranspileTraverseTests(TestCase):
 
     def test_multi_module_traversal(self):
         from render_static.tests.traverse import module1
-        from render_static.tests.traverse.module1 import (
-            Class1Module1,
-            Class2Module1,
-        )
+        from render_static.tests.traverse.module1 import Class1Module1, Class2Module1
         from render_static.tests.traverse.sub_pkg import module2
         from render_static.tests.traverse.sub_pkg.module2 import (
             Class1Module2,
@@ -493,10 +482,7 @@ class TranspileTraverseTests(TestCase):
         self.assertEqual(transpiler.TRAVERSAL, expected)
 
     def test_multi_class_traversal(self):
-        from render_static.tests.traverse.module1 import (
-            Class1Module1,
-            Class2Module1,
-        )
+        from render_static.tests.traverse.module1 import Class1Module1, Class2Module1
 
         transpiler = TranspilerTester()
         transpiler.transpile([Class1Module1, Class2Module1])
@@ -547,10 +533,7 @@ class TranspileTraverseTests(TestCase):
                 return ret
 
         from render_static.tests.traverse import module1
-        from render_static.tests.traverse.module1 import (
-            Class1Module1,
-            Class2Module1,
-        )
+        from render_static.tests.traverse.module1 import Class1Module1, Class2Module1
         from render_static.tests.traverse.sub_pkg.module2 import Class2Module2
 
         transpiler = EnumTranspiler()